+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.0.5
+
+* Fixed some known bugs.
+* Animation direction feature added. You can control sliding animation of your overlay popup dialog.
+* Added highlighting tapped child widget when overlay opened.
+* Added new playground example to see changes more precisely.
+
 ## 0.0.3
 
 * Web playground link added in README file.